Static::Strings!

+++ arangod/RestServer/Metrics.h
@@ -237,7 +237,7 @@ public:
     }
     _div = this->_delim.front();
     TRI_ASSERT(_div > T(0));
-    _lbase = logf(_base);
+    _lbase = log(_base);
   }
   virtual ~log_scale_t() {}
   /**
@@ -246,7 +246,7 @@ public:
    * @return    index
    */
   size_t pos(T const& val) const {
-    return static_cast<size_t>(1+std::floor(logf((val - this->_low)/_div)/_lbase));
+    return static_cast<size_t>(1+std::floor(log((val - this->_low)/_div)/_lbase));
   }
   /**
    * @brief Dump to builder
